+++ packages/code-du-travail-frontend/src/lib/matomo/index.ts
@@ -1,5 +1,2 @@
 export * from "./question";
 export * from "./types";
-export * from "./emitter";
-export * from "./indemnite-licenciement";
-export * from "./rupture-conventionnelle";
