+++ packages/code-du-travail-frontend/src/outils/CommonIndemniteDepart/events/emitter.ts
@@ -1,12 +1,4 @@
-export enum EventType {
-  SEND_RESULT_EVENT = "sendResultEvent",
-  TRACK_QUESTION = "trackQuestion",
-}
-
-export type CallbackEventType = {
-  [EventType.SEND_RESULT_EVENT]: (isEligible: boolean) => void;
-  [EventType.TRACK_QUESTION]: (title: string) => void;
-};
+import { CallbackEventType, EventType } from "./events";
 
 export const eventEmitter: {
   readonly events: Record<string, CallbackEventType[EventType][]>;
