+++ packages/code-du-travail-frontend/src/outils/CommonIndemniteDepart/steps/Resultat/store/store.ts
@@ -23,7 +23,8 @@ import { getInfoWarning } from "./service";
 import getSupportedCc from "../../../common/usecase/getSupportedCc";
 import * as Sentry from "@sentry/nextjs";
 import { CommonSituationStoreSlice } from "../../../../common/situationStore";
-import { EventType, eventEmitter } from "../../../../../lib";
+import { eventEmitter } from "../../../events/emitter";
+import { EventType } from "../../../events/events";
 
 const initialState: ResultStoreData = {
   input: {
